kata: cleanup. Moving all types to its own files. Preparing for Strategy pattern variation.
