+++ ConAppStrategyPattern/JsonExportService.cs
@@ -1,0 +1,9 @@
+﻿namespace ConAppStrategyPattern;
+
+public class JsonExportService : IExportSvc
+{
+	public void Export(Order order) 
+	{
+		Console.WriteLine($"Exporting {order.Name} to JSON");
+	}
+}
