+++ ConAppStrategyPattern/Order.cs
@@ -1,0 +1,9 @@
+﻿namespace ConAppStrategyPattern;
+
+public record Order(string Customer, int Amount, string Name, string Description, IExportSvc ExportService) 
+{
+	public void Export()
+	{
+		ExportService?.Export(this);
+	}
+}
